Schedule services, schedule algorithm fix
